GetCreditNote and update library

+++ app/src/main/java/co/dito/abako/apijitpack/data/model/request/delivery/SettlementDeliveryRequest.kt
@@ -3,7 +3,7 @@ package co.dito.abako.apijitpack.data.model.request.delivery
 import com.google.gson.annotations.SerializedName
 
 data class SettlementDeliveryRequest(
-    @SerializedName("Id") val deliveryId: Double,
+    @SerializedName("Id") val deliveryId: Int,
     @SerializedName("DllLiq") val settlementDetail: List<SettlementDetailDeliveryRequest>,
     @SerializedName("DllOc") val otherCharges: List<OtherChargesDetail>
 )
