+++ app/src/main/java/co/dito/abako/apijitpack/data/model/request/delivery/SettlementDetailDeliveryRequest.kt
@@ -3,6 +3,6 @@ package co.dito.abako.apijitpack.data.model.request.delivery
 import com.google.gson.annotations.SerializedName
 
 data class SettlementDetailDeliveryRequest(
-    @SerializedName("IdTv") val idValueType: Double,
+    @SerializedName("IdTv") val idValueType: Int,
     @SerializedName("VC") val valueQuantity: Int,
 )
